+++ 9space_snapshot_api/constants.py
@@ -1,0 +1,10 @@
+"""Fixed network contract for the site-local Dahua producer."""
+
+NVR_RTSP_PORT = 554
+NVR_HTTP_PORT = 80
+HUB_TELEMETRY_PORT = 8765
+SNAPSHOT_HOST_PORT = 8222
+SNAPSHOT_CONTAINER_PORT = 8000
+MAGICDNS_SERVER = "100.100.100.100"
+LOCAL_HUB_HOSTNAME = "afa94ae2-9space-monitor-hub"
+LOCAL_SNAPSHOT_HOSTNAME = "afa94ae2-9space-snapshot-addon"